Save as draft

+++ article/models/models.py
@@ -8,11 +8,12 @@ from article.signoffs import publish_article_signoff
 
 
 class Article(models.Model):
-    title = models.CharField(max_length=200,)
+    title = models.CharField(max_length=200, null=False, blank=False)
     author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='article_author')
     summary = models.TextField(max_length=100, null=False, blank=False)
-    article_text = models.TextField(max_length=1000,)
+    article_text = models.TextField(max_length=1000, null=False, blank=False)
 
+    is_published = models.BooleanField(default=False)
     publish_signoff, publish_signet = SignoffField(publish_article_signoff)
 
     likes = SignoffSet('like_signoff')
@@ -24,8 +25,9 @@ class Article(models.Model):
             return f"{self.author.username} - {self.title}"
 
     def delete(self, *args, **kwargs):
-        self.publish_signet.delete()     # Delete the signet associated with the article
-        super().delete(*args, **kwargs)  # Delete the article itself
+        if self.is_published:
+            self.publish_signet.delete()  # Delete the signet associated with the article
+        super().delete(*args, **kwargs)   # Delete the article itself
 
     def total_likes(self):
         return self.likes.count()
@@ -40,6 +42,16 @@ class Article(models.Model):
             return self.author.get_full_name()
         else:
             return self.author.username
+
+    def publish(self, user):
+        self.is_published = True
+        self.save()
+        self.publish_signoff.sign_if_permitted(user)
+
+    def unpublish(self, user):
+        self.is_published = False
+        self.save()
+        self.publish_signoff.revoke_if_permitted(user)
 
 
 class LikeSignet(Signet):
@@ -64,7 +76,7 @@ class Comment(models.Model):
 
 
 class CommentSignet(Signet):
-    # ForeignKey(unique=True) is usually better served by a OneToOneField
+    # TODO: Replace ForeignKey with SignoffUnique
     comment = models.ForeignKey(Comment, unique=True, on_delete=models.CASCADE, related_name='signatories')
